Adapt the sigmoid_focal_loss operator on the parrots camb device for mmcv
