+++ mmcv/ops/csrc/parrots/camb/focal_loss_camb.mlu
@@ -10,7 +10,6 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *************************************************************************/
 #include <float.h>
-#include "bang_internal.h"
 
 #define NFU_ALIGN_SIZE 128          // Byte
 #define REM_FOR_STACK (128 * 1024)  // 128KB reserved for cncc
